+++ define/oauth-kakao.ts
@@ -1,0 +1,2 @@
+// oauth kakao
+
