+++ view_model/index.tsx
@@ -1,0 +1,3 @@
+import {Test} from "../view_model/vm-oauth-kakao";
+
+export {Text}
